loader, interval 1m

+++ client/src/store.js
@@ -36,18 +36,28 @@ export default create((set, get) => ({
       set({ timeline: cachedTimeline, since })
     }
   },
-  async getUserTimeline () {
+  async getUserTimeline ({ force = false } = {}) {
+    set({ loading: true })
+    if (force) {
+      return ApiService.getUserTimeline()
+        .then(timeline => {
+          const since = timeline.reduce((newest, curr) => newest < +new Date(curr.time) ? +new Date(curr.time) : newest, +new Date(timeline[0].date))
+          set({ timeline, since, loading: false })
+          window.localStorage.setItem('timeline', JSON.stringify(timeline))
+        })
+        .catch(err => {
+          console.error(err)
+          set({ timeline: [], loading: false })
+        })
+    }
     const since = get().since
     const cachedTimeline = get().timeline
-    set({ loading: true })
-    ApiService.getUserTimeline({ since })
-      .then(newTimeline => {
-        const timeline = newTimeline.map(t => Object.assign(t, { status: '/status' + t.link.replace(/.*\/status/, '').replace(/#.*/, '') }))
+    return ApiService.getUserTimeline({ since })
+      .then(timeline => {
         if (cachedTimeline && cachedTimeline.length > 0) timeline.push(...cachedTimeline)
         const since = timeline.reduce((newest, curr) => newest < +new Date(curr.time) ? +new Date(curr.time) : newest, +new Date(timeline[0].date))
         set({ timeline, since, loading: false })
         window.localStorage.setItem('timeline', JSON.stringify(timeline))
-        setTimeout(() => { set({ loading: false }) }, 1500)
       })
       .catch(err => {
         console.error(err)
